refactor: split miniscript and descriptor into separate files
Issue: BTC-1348
